Moved and consoladated printing methods to a new file, scan_manager. Also moved scan manager related methods to this new module. Refactored db to be more efficient by just retrieving ids instead of the entire JSON blob
